+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Heap.shared.resetIdentity()` and `Heap.shared.identify()` no longer clear event properties by default when a new user is identified.
   The previous behavior is available using the option `.clearEventPropertiesOnNewUser`.
 
+- Changed uploader behavior around server errors.
+
 ### Added
 
 - Added option `.clearEventPropertiesOnNewUser` to continue using existing SDK behavior where event properties are cleared when a new user is identified.